+++ src/integrationTest/java/uk/gov/hmcts/divorce/document/content/GeneralOrderTemplateContentIT.java
@@ -45,7 +45,7 @@ public class GeneralOrderTemplateContentIT {
 
         var ctscContactDetails = CtscContactDetails
             .builder()
-            .centreName("HMCTS Digital Divorce")
+            .centreName("HMCTS Digital Divorce and Dissolution")
             .emailAddress("divorcecase@justice.gov.uk")
             .serviceCentre("Courts and Tribunals Service Centre")
             .poBox("PO Box 12706")
