+++ src/main/java/uk/gov/hmcts/divorce/divorcecase/model/CtscContactDetails.java
@@ -3,10 +3,12 @@ package uk.gov.hmcts.divorce.divorcecase.model;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Builder;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 
 @Data
 @Builder
 @JsonIgnoreProperties(ignoreUnknown = true)
+@EqualsAndHashCode
 public class CtscContactDetails {
     private String serviceCentre;
