+++ src/main/java/uk/gov/hmcts/divorce/solicitor/event/SolicitorSubmitJointApplication.java
@@ -18,6 +18,7 @@ import uk.gov.hmcts.divorce.solicitor.event.page.CheckApplicant1SolicitorAnswers
 import uk.gov.hmcts.divorce.solicitor.event.page.FinancialOrdersForApplicant2;
 import uk.gov.hmcts.divorce.solicitor.event.page.HelpWithFeesPageForApplicant2;
 import uk.gov.hmcts.divorce.solicitor.event.page.MarriageIrretrievablyBrokenForApplicant2;
+import uk.gov.hmcts.divorce.solicitor.service.SolicitorSubmitApplicationService;
 import uk.gov.hmcts.divorce.systemupdate.service.CcdUpdateService;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
 import uk.gov.hmcts.reform.ccd.client.model.SubmittedCallbackResponse;
@@ -54,16 +55,7 @@ public class SolicitorSubmitJointApplication implements CCDConfig<CaseData, Stat
     private HelpWithFeesPageForApplicant2  helpWithFeesPageForApplicant2;
 
     @Autowired
-    private CcdUpdateService ccdUpdateService;
-
-    @Autowired
-    private IdamService idamService;
-
-    @Autowired
-    private AuthTokenGenerator authTokenGenerator;
-
-    @Autowired
-    private HttpServletRequest httpServletRequest;
+    private SolicitorSubmitApplicationService solicitorSubmitApplicationService;
 
     @Override
     public void configure(final ConfigBuilder<CaseData, State, UserRole> configBuilder) {
@@ -101,29 +93,9 @@ public class SolicitorSubmitJointApplication implements CCDConfig<CaseData, Stat
 
         log.info("Solicitor submit joint application submitted callback invoked for case id: {}", details.getId());
 
-        submitEventForApprovalOrRequestingChanges(details);
+        solicitorSubmitApplicationService.submitEventForApprovalOrRequestingChanges(details);
 
         return SubmittedCallbackResponse.builder().build();
     }
 
-    @Async
-    public void submitEventForApprovalOrRequestingChanges(CaseDetails<CaseData, State> details) {
-        final Application application = details.getData().getApplication();
-
-        String eventId;
-        if (YES.equals(application.getApplicant2ConfirmApplicant1Information())) {
-            eventId = CITIZEN_APPLICANT_2_REQUEST_CHANGES;
-        } else {
-            eventId = APPLICANT_2_APPROVE;
-        }
-
-        String solicitorAuthToken = httpServletRequest.getHeader(AUTHORIZATION);
-        User solUser = idamService.retrieveUser(solicitorAuthToken);
-
-        final String serviceAuthorization = authTokenGenerator.generate();
-
-        log.info("Submitting event id {} for case id: {}", eventId, details.getId());
-
-        ccdUpdateService.submitEvent(details, eventId, solUser, serviceAuthorization);
-    }
 }
