+++ src/main/java/uk/gov/hmcts/divorce/solicitor/service/SolicitorSubmitApplicationService.java
@@ -1,26 +1,63 @@
 package uk.gov.hmcts.divorce.solicitor.service;
 
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
+import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
 import uk.gov.hmcts.ccd.sdk.type.OrderSummary;
+import uk.gov.hmcts.divorce.divorcecase.model.Application;
+import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
+import uk.gov.hmcts.divorce.divorcecase.model.State;
+import uk.gov.hmcts.divorce.idam.IdamService;
 import uk.gov.hmcts.divorce.payment.model.Payment;
+import uk.gov.hmcts.divorce.systemupdate.service.CcdUpdateService;
+import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
+import uk.gov.hmcts.reform.idam.client.models.User;
+
+import javax.servlet.http.HttpServletRequest;
 
 import static java.lang.Integer.parseInt;
+import static org.springframework.http.HttpHeaders.AUTHORIZATION;
+import static uk.gov.hmcts.ccd.sdk.type.YesOrNo.YES;
+import static uk.gov.hmcts.divorce.citizen.event.CitizenApplicant2Approve.APPLICANT_2_APPROVE;
+import static uk.gov.hmcts.divorce.citizen.event.CitizenApplicant2RequestChanges.CITIZEN_APPLICANT_2_REQUEST_CHANGES;
 import static uk.gov.hmcts.divorce.payment.model.PaymentStatus.SUCCESS;
 
 @Service
 @Slf4j
 public class SolicitorSubmitApplicationService {
 
-    public Payment getDummyPayment(final OrderSummary orderSummary) {
-        return Payment
-            .builder()
-            .amount(parseInt(orderSummary.getPaymentTotal()))
-            .channel("online")
-            .feeCode("FEE0001")
-            .reference(orderSummary.getPaymentReference())
-            .status(SUCCESS)
-            .transactionId("ge7po9h5bhbtbd466424src9tk")
-            .build();
+    @Autowired
+    private CcdUpdateService ccdUpdateService;
+
+    @Autowired
+    private IdamService idamService;
+
+    @Autowired
+    private AuthTokenGenerator authTokenGenerator;
+
+    @Autowired
+    private HttpServletRequest httpServletRequest;
+
+    @Async
+    public void submitEventForApprovalOrRequestingChanges(CaseDetails<CaseData, State> details) {
+        final Application application = details.getData().getApplication();
+
+        String eventId;
+        if (YES.equals(application.getApplicant2ConfirmApplicant1Information())) {
+            eventId = CITIZEN_APPLICANT_2_REQUEST_CHANGES;
+        } else {
+            eventId = APPLICANT_2_APPROVE;
+        }
+
+        String solicitorAuthToken = httpServletRequest.getHeader(AUTHORIZATION);
+        User solUser = idamService.retrieveUser(solicitorAuthToken);
+
+        final String serviceAuthorization = authTokenGenerator.generate();
+
+        log.info("Submitting event id {} for case id: {}", eventId, details.getId());
+
+        ccdUpdateService.submitEvent(details, eventId, solUser, serviceAuthorization);
     }
 }
